Shut up compiler warnings.

+++ chocolate-doom/src/p_enemy.c
@@ -420,7 +420,7 @@ void P_NewChaseDir (mobj_t*	actor)
 	&& d[2] != DI_NODIR)
     {
 	actor->movedir = diags[((deltay<0)<<1)+(deltax>0)];
-	if (actor->movedir != turnaround && P_TryWalk(actor))
+	if (actor->movedir != (int) turnaround && P_TryWalk(actor))
 	    return;
     }
 
@@ -473,7 +473,7 @@ void P_NewChaseDir (mobj_t*	actor)
 	      tdir<=DI_SOUTHEAST;
 	      tdir++ )
 	{
-	    if (tdir!=turnaround)
+	    if (tdir != (int) turnaround)
 	    {
 		actor->movedir =tdir;
 		
@@ -488,9 +488,9 @@ void P_NewChaseDir (mobj_t*	actor)
 	      tdir != (DI_EAST-1);
 	      tdir-- )
 	{
-	    if (tdir!=turnaround)
+	    if (tdir != (int) turnaround)
 	    {
-		actor->movedir =tdir;
+		actor->movedir = tdir;
 		
 		if ( P_TryWalk(actor) )
 		    return;
